Introduce static methods JsonValue.readFrom
New methods are:
  JsonValue.readFrom( String ) -> JsonValue
  JsonValue.readFrom( Reader ) -> JsonValue

  JsonObject.readFrom( String ) -> JsonObject
  JsonObject.readFrom( Reader ) -> JsonObject

  JsonArray.readFrom( String ) -> JsonObject
  JsonArray.readFrom( Reader ) -> JsonObject

Using these methods, public access to JsonParser is not required
anymore to parse JSON, so make JsonParser package private.

Get rid of the distinction between parsing a value or an array / object.
JsonParser.parse() accepts all kinds of values. If an array or an object
is expected, the readFrom methods on JsonObject/JsonArray can be used.