Improve object lookup performance by adding a small hash index
Typical JSON messages include objects with only a few members.
Looking up object members can be improved by storing the hash of
a name in a map that points to the index in the names list.

For small item counts, an array of 32 bytes provides the best
trade-off between footprint, hashing overhead, and lookup time.

Creating, filling and looking up n members is
* almost the same as a plain ArrayList for item counts < 4
* much faster than ArrayList for 4..32 items
* faster than a HashMap for item counts < 30
For bigger item counts, the impact disappears.