Review how Buffered Readers and Writers affect performance
Benchmarks show that wrapping an existing Reader in an additional
BufferedReader does not improve reading performance, as the parser
already uses a buffer internally.

In contrast to that, using a BufferedWriter can dramatically improve
writing performance. However, the JsonWriter should not wrap the given
writer in a BufferedWriter itself, as this would add a significant
overhead when the given writer is already a BufferedWriter or a
StringWriter.

Add respective advice to JavaDoc of readFrom(Reader) and writeTo(Writer)
methods.