Integrate buffered reading code into JsonParser
By moving the buffered reading code into the JsonParser it can be
combined with the use of the capture buffer. Instead of shifting or
extending the buffer while capturing, the captured region is now simply
copied into the capture buffer on refill. This simplifies the code and
spares one class.

Performance tests show a slight improvement for inputs hat contain long
strings.