Benchmarks: writeToWriter should not do UTF-8 conversion.
This makes it consistent with timeReadFromReader, which reads from a
String (and not from UTF-8 bytes).

The new timeWriteToOutputStream covers the case of writing to UTF-8
bytes.

But now there's a question of whether this test useful, since it seems
unlikely for someone to use a Writer without ultimately ending up
writing to UTF-8 bytes.