Benchmarks: Remove unneeded buffering from ReadWriteBenchmark
As suggested in #38.

Both ByteArrayOutputStream and StringWriter contain a memory buffer,
there is no need to wrap them with another buffer.

Buffering helps when it is more efficient to make fewer large writes
than it is to make many small writes (e.g. the 'write' syscall).  But
since these benchmarks just use in-memory buffers for the final
destination, buffering is not useful; it just adds noise to the results.