Benchmarks: Remove buffering from writeToWriter/writeToOutputStream.
Buffering helps when it is more efficient to make fewer large writes
than it is to make many small writes (e.g. the 'write' syscall).  But
since these benchmarks just use in-memory buffers for the final
destination, buffering is not useful; it just adds noise to the results.

However, it is useful to wrap buffering around OutputStreamWriter, and
so I did that in JsonRunner.writeToOutputStream.