Download d3 lib in maven build
Add a build step that fetches the d3 js file and place it in lib/.
This makes it easier to get started without having to include the lib
in this repo. Once fetched, the file doesn't have to be fetched again
and is also available offline.

Print an error message if d3 is missing.