Benchmarks: use buffered writer in writeToOutputStream
Since minimal-json writes small chunks directy to the Writer, using a
buffered writer drastically increases writing performance (as shown in
BufferedWriterBenchmark).

Wrap the writer in the default implementation of writeToOutputStream().
For parsers that don't provide a method to write to an OutputStream,
we want to compare with the best writing performance that can be
achived with a Writer.