Add benchmark for double buffering of a writer
I wanted to understand what the overhead of wrapping a buffered writer
in another buffered writer is. Turns out that it's somewhere in the
range of microseconds and almost neglectable in most cases.

However, I still don't like to wrap the writer in a buffered writer
because it might have a negative impact when writing a large number
of small JSON documents to an already buffered writer.