Introduce small writing buffer to improve performance
Reduce the amount of writing operations called on the given writer. This
improves writing performance even when writing to a StringWriter. As an
effect, it also speeds up toString().

The default size of 128 chars turned out to be a good compromise. It
hardly affects the performance for smaller documents, but sigificantly
speeds up writing bigger ones.