Rework JsonLiteral, fix serializable issue
The methods isTrue(), isBoolean() etc. did not work on deserialized
literals, because they checked for object identity. Instead of an
implementation that checks for equality, keep the results in boolean
fields that in the constants.

There can't be other literal constants than `null`, `true`, and `false`
in JSON. Make JsonLiteral constructor private and expose only the three
constant values.