Don't accept null as WriterConfig parameter, improve tests
Calling JsonValue.toString() and JsonValue.writeTo() with a null config
should not be supported since

- WriterConfig.MINIMAL exists and has the same effect
- the config parameter is optional in both methods

Replace the use of null with WriterConfig.MINIMAL and improve tests for
NPEs including checks for exception messages.

Allow calling TestUtil.assertException with methods that throw checked
exceptions by introducing an interface RunnableEx that is similar to
Runnable but supports Exceptions thrown by the run method.