Introduce util class Json for constants and static methods
Since JsonValue has two subclasses JsonArray and JsonObject, the
constants NULL, TRUE, and FALSE defined on JsonValue are available
on the subclasses by inheritance. For example, the expression

    JsonArray.TRUE

actually compiles even though it's non-sense. The same is true for
the static factory methods. I've seen this happen, despite compiler
warnings, and it can result in pretty incomprehensible code.

Move constants to a dedicated util class Json and deprecate the
constants on JsonValue to prevent those cases.

This also results in shorter code and matches common language: for
example, Json.NULL represents "json null".

In a next step, static factory methods can be moved to this class as
well.

See #52