Move static factory methods from JsonValue to Json
Same as with constants, the static methods on JsonValue are available
by the subclasses JsonArray and JsonObject by inheritance. This allows
for expressions like

    JsonArray.valueOf(23)

which would not return a JsonArray, as you could expect, but a
JsonValue.

Move static factory methods to the new util class Json and deprecate
the methods on JsonValue to prevent those cases.

This also results in shorter code:

    Json.value(23)

instead of

    JsonValue.valueOf(23)

Fix #52