Provide static factory methods for JsonArray
There is no convenient method to convert Java arrays to a JsonArrays.
Array elements have to be added to a new JsonArray in a loop like this:

    JsonArray json = new JsonArray();
    for (String name : names) {
      json.add(name);
    }

Introduce static factory methods on the utility class Json to create
JsonArrays from a given varargs parameter. Using these methods, the
code above can be written as a one-liner:

    JsonArray json = Json.array(names);

Since the methods support varargs, they can also replace this:

    new JsonArray().add(23).add(42)

with:

    Json.array(23, 42)

Also introduce a parameterless method Json.array() as a shortcut for
the JsonArray constructor. This makes the API more consistent and
prevents "ambiguous method" compiler errors when Json.array() is
called without parameters.

Fix #12