Introduce parse methods on Json class
Make the relevant methods easier to find by using the Json class as
the central entry point to the minimal-json API. Use this class not
only to create JSON structures, but also to parse JSON.

Instead of using the static methods on JsonValue, JsonArray, and
JsonObject:

    JsonValue value = JsonValue.readFrom(input);
    JsonArray array = JsonArray.readFrom(input);
    JsonObject object = JsonObject.readFrom(input);

provide Json.parse() methods to be used with asXxx():

    JsonValue value = Json.parse(input);
    JsonArray array = Json.parse(input).asArray();
    JsonObject object = Json.parse(input).asObject();

For arrays and objects, this results in two more characters, however,
the new syntax seems to fit better with the rest of the API. Same as
with the get() methods on JsonArray and JsonObject, the parse() methods
return a JsonValue that can be transferred to the expected type using
asXxx() methods.

Mark the static readFrom() methods on JsonValue, JsonArray, and
JsonObject as deprecated.