Introduce Streaming API
Introduce base class JsonHandler with methods to be called on every
parser event. Add a method to obtain the current location as an
immutable object to allow using it at a later point.

Let JsonParser issue events to a given JsonHandler instead of creating
a JsonValue itself. Use an internal JsonHandler to create JsonValues.

Publish JsonParser API. This class has no reference to JsonValue
anymore. It can be used to create arbitrary data structures.

Fix #17