Fix offset off-by-one at end of input
The read() method fails to increment the index when reading the last
character of the input. Other methods compensate for this by treating
the index differently when at end of text.

By correctly incrementing the index also when at end of text, the
conditional adjustments can be removed from endCapture() and
getLocation().

Make unit tests for the parse() method more comprehensible by checking
the offset only. The calculation of line and column should be tested
individually.