Extract inner class Location, use it in ParseException
ParseException offers the error location as individual fields offset,
line, and column, whereas JsonHandler has to provide as a result of a
method call, i.e. as an object.

Provide location consistently as Location object in both places.
Deprecate fields in ParseException.

Extract inner class Location from JsonParser as it is now referred to
by JsonHandler and ParseException.