Remove deprecated API
As reported in #77, the reference from the static fields in `JsonValue`
to static fields in `Json` can lead to deadlocks during class
initialization.

Since this API has only been added for compatibility, the easiest fix
seems to be to remove the deprecated fields in `JsonValue`.

Use this opportunity to remove all deprecated API for 0.9.5.

This commit should fix #77.